// projects/cooee-admin/src/app/models/organization/stream.enum.ts
export enum Stream {

    CIVIL_ENGINEERING = 'CIVIL_ENGINEERING',
    MECHANICAL_ENGINEERING = 'MECHANICAL_ENGINEERING',
    BBA = 'BBA',
    AGRICULTURE = 'AGRICULTURE',
    ENVIRONMENTAL_SCIENCE = 'ENVIRONMENTAL_SCIENCE',
    COMPUTER_ENGINEERING = 'COMPUTER_ENGINEERING',
    ELECTRICAL_ENGINEERING = 'ELECTRICAL_ENGINEERING',
    ELECTRONICS_ENGINEERING = 'ELECTRONICS_ENGINEERING',
    INFORMATION_TECHNOLOGY = 'INFORMATION_TECHNOLOGY',
    PHYSICAL_SCIENCES = 'PHYSICAL_SCIENCES',
    ECONOMICS = 'ECONOMICS',
    HISTORY = 'HISTORY',
    POLITICAL_SCIENCE = 'POLITICAL_SCIENCE',
    PHILOSOPHY = 'PHILOSOPHY',
    PSYCHOLOGY = 'PSYCHOLOGY',
    SOCIOLOGY_AND_COMMERCE = 'SOCIOLOGY_AND_COMMERCE',
    OTHER = 'OTHER'
}